ML/requirements.txt
paddleocr==2.8.1
paddlepaddle==2.6.2
Pillow==10.4.0
numpy==1.26.4
opencv-python==4.10.0.84
